+++ packages/core/src/storage/extract.ts
@@ -1,18 +1,20 @@
-import assert from 'assert';
+import assert from 'assert/strict';
 import {
   ContractDefinition,
   StructDefinition,
   EnumDefinition,
   TypeDescriptions,
   VariableDeclaration,
+  TypeName,
 } from 'solidity-ast';
 import { isNodeType, findAll, ASTDereferencer } from 'solidity-ast/utils';
-import { StorageLayout, TypeItem } from './layout';
+import { StorageLayout, StructMember, TypeItem, isStructMembers, EnumMember } from './layout';
 import { normalizeTypeIdentifier } from '../utils/type-id';
 import { SrcDecoder } from '../src-decoder';
 import { mapValues } from '../utils/map-values';
 import { pick } from '../utils/pick';
 import { execall } from '../utils/execall';
+import { loadNamespaces } from './namespace';
 
 const currentLayoutVersion = '1.2';
 
@@ -20,31 +22,53 @@ export function isCurrentLayoutVersion(layout: StorageLayout): boolean {
   return layout?.layoutVersion === currentLayoutVersion;
 }
 
+export interface CompilationContext {
+  deref: ASTDereferencer;
+  contractDef: ContractDefinition;
+  storageLayout?: StorageLayout;
+}
+
 export function extractStorageLayout(
   contractDef: ContractDefinition,
   decodeSrc: SrcDecoder,
   deref: ASTDereferencer,
-  storageLayout?: StorageLayout | undefined,
+  storageLayout?: StorageLayout,
+  namespacedContext?: CompilationContext,
 ): StorageLayout {
   const layout: StorageLayout = { storage: [], types: {}, layoutVersion: currentLayoutVersion, flat: false };
-  if (storageLayout !== undefined) {
-    layout.types = mapValues(storageLayout.types, m => {
-      return {
-        label: m.label,
-        members: m.members?.map(m =>
-          typeof m === 'string' ? m : pick(m, ['label', 'type', 'offset', 'slot']),
-        ) as TypeItem['members'],
-        numberOfBytes: m.numberOfBytes,
-      };
-    });
 
+  // The namespaced context will contain the types of namespaces that may not be included
+  // in the original storage layout.
+  // Some types will be present in both and they must be exactly equivalent.
+  // If they are not, it fails an assertion because this may be a clash between different types.
+  const combinedTypes = { ...namespacedContext?.storageLayout?.types, ...storageLayout?.types };
+  if (namespacedContext?.storageLayout?.types) {
+    for (const t in storageLayout?.types) {
+      if (t in namespacedContext.storageLayout.types) {
+        assert.deepEqual(namespacedContext.storageLayout.types[t], storageLayout?.types[t]);
+      }
+    }
+  }
+
+  layout.types = mapValues(combinedTypes, m => {
+    return {
+      label: m.label,
+      members:
+        m.members && isStructMembers(m.members)
+          ? m.members.map(m => pick(m, ['label', 'type', 'offset', 'slot']))
+          : m.members,
+      numberOfBytes: m.numberOfBytes,
+    };
+  });
+
+  if (storageLayout !== undefined) {
     for (const storage of storageLayout.storage) {
       const origin = getOriginContract(contractDef, storage.astId, deref);
       assert(origin, `Did not find variable declaration node for '${storage.label}'`);
       const { varDecl, contract } = origin;
       const { renamedFrom, retypedFrom } = getRetypedRenamed(varDecl);
       // Solc layout doesn't bring members for enums so we get them using the ast method
-      loadLayoutType(varDecl, layout, deref);
+      loadLayoutType(varDecl.typeName, layout, deref);
       const { label, offset, slot, type } = storage;
       const src = decodeSrc(varDecl);
       layout.storage.push({ label, offset, slot, type, contract, src, retypedFrom, renamedFrom });
@@ -65,11 +89,15 @@ export function extractStorageLayout(
             renamedFrom,
           });
 
-          loadLayoutType(varDecl, layout, deref);
+          loadLayoutType(varDecl.typeName, layout, deref);
         }
       }
     }
   }
+
+  const origContext = { deref, contractDef, storageLayout };
+  loadNamespaces(decodeSrc, layout, origContext, namespacedContext);
+
   return layout;
 }
 
@@ -92,14 +120,34 @@ function typeDescriptions(x: { typeDescriptions: TypeDescriptions }): RequiredTy
   return x.typeDescriptions as RequiredTypeDescriptions;
 }
 
-function getTypeMembers(typeDef: StructDefinition | EnumDefinition): TypeItem['members'] {
+type GotTypeMembers<
+  D extends EnumDefinition | StructDefinition,
+  F extends 'src' | 'typeName',
+> = D extends EnumDefinition ? EnumMember[] : (StructMember & Pick<VariableDeclaration, F>)[];
+
+export function getTypeMembers<D extends EnumDefinition | StructDefinition>(typeDef: D): GotTypeMembers<D, never>;
+export function getTypeMembers<D extends EnumDefinition | StructDefinition, F extends 'src' | 'typeName'>(
+  typeDef: D,
+  includeFields: { [f in F]: true },
+): GotTypeMembers<D, F>;
+export function getTypeMembers(
+  typeDef: StructDefinition | EnumDefinition,
+  includeFields: { src?: boolean; typeName?: boolean } = {},
+): TypeItem['members'] {
   if (typeDef.nodeType === 'StructDefinition') {
     return typeDef.members.map(m => {
       assert(typeof m.typeDescriptions.typeIdentifier === 'string');
-      return {
+      const member: StructMember & Partial<Pick<VariableDeclaration, 'src' | 'typeName'>> = {
         label: m.name,
         type: normalizeTypeIdentifier(m.typeDescriptions.typeIdentifier),
       };
+      if (includeFields.src && m.src) {
+        member.src = m.src;
+      }
+      if (includeFields.typeName && m.typeName) {
+        member.typeName = m.typeName;
+      }
+      return member;
     });
   } else {
     return typeDef.members.map(m => m.name);
@@ -116,16 +164,16 @@ function getOriginContract(contract: ContractDefinition, astId: number | undefin
   }
 }
 
-function loadLayoutType(varDecl: VariableDeclaration, layout: StorageLayout, deref: ASTDereferencer) {
+export function loadLayoutType(typeName: TypeName | null | undefined, layout: StorageLayout, deref: ASTDereferencer) {
   // Note: A UserDefinedTypeName can also refer to a ContractDefinition but we won't care about those.
   const derefUserDefinedType = deref(['StructDefinition', 'EnumDefinition', 'UserDefinedValueTypeDefinition']);
 
-  assert(varDecl.typeName != null);
+  assert(typeName != null);
 
   // We will recursively look for all types involved in this variable declaration in order to store their type
   // information. We iterate over a Map that is indexed by typeIdentifier to ensure we visit each type only once.
   // Note that there can be recursive types.
-  const typeNames = new Map([...findTypeNames(varDecl.typeName)].map(n => [typeDescriptions(n).typeIdentifier, n]));
+  const typeNames = new Map([...findTypeNames(typeName)].map(n => [typeDescriptions(n).typeIdentifier, n]));
 
   for (const typeName of typeNames.values()) {
     const { typeIdentifier, typeString: label } = typeDescriptions(typeName);